Add delete workspace validation

+++ src/main/java/com/example/back/backend/application/service/WORKSPACE/WSP0400ASVC.java
@@ -1,8 +1,12 @@
 package com.example.back.backend.application.service.WORKSPACE;
 
+import com.example.back.backend.Exception.BizException;
 import com.example.back.backend.application.dto.workspace.WSP0400AInput;
 import com.example.back.backend.application.dto.workspace.WSP0400AOutput;
+import com.example.back.backend.common.Enum.AccountEnum;
+import com.example.back.backend.common.Enum.SysErrCode;
 import com.example.back.backend.infrastructure.entity.PK.WorkspaceInfoPk;
+import com.example.back.backend.infrastructure.entity.WorkspaceInfo;
 import com.example.back.backend.infrastructure.jpa.DaoWorkspaceInfoJpa;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -38,6 +42,12 @@ public class WSP0400ASVC {
         WorkspaceInfoPk workspaceInfoPk = new WorkspaceInfoPk();
         workspaceInfoPk.setWorkspaceId(ctxSVC.input.getWorkspaceId());
         workspaceInfoPk.setAdminId(ctxSVC.input.getAdminId());
+
+        WorkspaceInfo workspaceInfo = workspaceInfoJpa.findById(workspaceInfoPk).orElseThrow(() -> new Exception("Data Not Found"));
+
+        if (workspaceInfo.getHierarchy().equals(AccountEnum.workspace.OWNER.getValue())){
+            throw new BizException(SysErrCode.WORKSPACE_ADMIN_DEL_ERR);
+        }
 
         workspaceInfoJpa.deleteById(workspaceInfoPk);
